Added a top level script to run the pipeline
- The script gireds.py is to be the top level interface between the
various data manipulation functions and the user, along with the
central hub for directory structuring and file association. It is
designed to work with a comprehensive configuration file through
the ConfigParser module.

- example.cfg is an example of such a configuration file, containing
directory definitions (at this point), options for the reduction
process, and basically anything we would want the final user to be able
to change.