Changes in the sky subtraction
During the re-reduction of NGC 3982 I came accross a curious step in
the velocity field that led me to investigate the matter more closely.
It turns out that Carine Brum had recently published a paper with the
same data from 2007 and therefore I could check if the problem was with
the procedures in GIREDS. Their results, however, were equal to mine,
and although we think it is probably an instrumental artifact I could
not find the source of the problem. Actually the sky lines appear
shifted in one of the slits by about 0.3 pixels.

Alas, when one looks carefully into a problem one usually finds more. I
noticed that, as it happens with some other galaxies, NGC 3982 had a
lot of spectra with negative flux values. That is, the spectra was
shifted by an additive term, rather than reflected in the flux
direction. The usual suspect for this kind of problem is the sky
subtraction, since it is the only additive step other than the BIAS.
However, upon checking the sky sampling fibers, I found that many of
them seemed to be inexplicably getting brighter towards the aperture
numbers 750-ish.

This seemed to be caused by over-correction by the fiber response map,
and thus I chased the culprit down that alley. After many tests and
re-reductions, including some interactive aperture identification, I
realized the fiber response was mostly correct, with only a few minor
inaccuracies that have already been corrected in a previous commit. The
problem was that when the response was applied to a mainly constant
flux accross the sky fibers, the dimmer of then were greatly
multiplied.

Finally, I realized the solution was to be found in the GFSKYSUB
routine, and the way it was evaluating the average sky spectrum. The
previous behaviour was the standard averaging of the fibers in both
slits. This first problem was promptly identifed and corrected, since
the new versions of the Gemini package has an option to produce a sky
spectrum for each slit.

The main problem was with the combination procedure. Averaging over all
the fibers gave too much weight to those over-corrected sky fibers at
the edge of the pseudo-slit. My solution, which has produced
satisfactory results, was to change the combination mode to MEDIAN and
the sigma limits for rejection to 1. After that, only a couple of
pixels ended up negative.