Finding dead fibers
Added a function to identify dead fibers in the 50 fiber bundles.
Surprisingly enough it works for the few tests I ran. The basic idea
is to produce a model of the fiber positions, given the first fiber
position and the distance between adjacent fibers. Dead fibers at the
beginning, middle and end of the bundles have been correctly
identified. The next step should be to write these results to the MDF
and see if GFEXTRACT gets it right.