Attempt at generic first and last fibers fixer
The function find_dead_beams has been severely modified in an attempt
to make a generic fixer for missing fibers at the beginning and end of
the bundle. The current form does not work when the next bundle has a
missing first fiber. A possible solution is to check if the same is
true for the next iteration, that is, see if the next bundle also has a
missing first fiber. But that would cause problems when there are
missing first and last fibers.

I decided to change completely my approach and made a new function to
read the apertures from the database file. This will serve as the
starting point for a new method of identifying dead fibers.