Automatic identification of fiber bundles
Arguably the most time consuming step in the reduction of IFU data
from GMOS is the correct identification of fiber bundles on the flat
field image. This step is crucial, as it defines where each spectrum
will be placed in the final data cube.

GIREDS now has a function which performs the identification of dead
fibers automatically, and also checks that the correct solution was
also found by GFEXTRACT.