Working commit of cube builder
This commit achieves a working version of the cube builder, which has
the correct treatment for all the extensions: SCI, VAR and DQ.

Additionaly, an atmospheric refraction correction is implemented,
based on the observed shift in position of the source. This will not
work if the data cube lacks a clear peak in flux.