Severe changes in the atmospheric refraction correction
After a few unsuccessful attempts at obtaining the correct fix for the
atmospheric refraction in the NGC~4180 data cube I decided to rewrite
part of the code.

The problem with this galaxy is mainly that the brightest point in each
image is not at the center, but rather at the image's edge. Since the
previous version relied on finding the center of mass of each image,
the results were far from accurate.

After the rewrite the code is much more readable and streamlined.
The main changes are as follows:
- Offsets are obtained via the image_registration.chi2_shift function.
- Offsets are based on a "reference" image, which is actually the median
  along the wavelength axis.
- Each plane (segments of the data cube collapsed along the wavelength
  direction) is independently normalized to its maximum value, so that
  offsets are independent of the spectrum shape.
- Removed sigma clipping from the offset fitting.
- Added a debug option for plotting the residuals between data and
  shifted planes.
- Added a parameter in fit_refraction_function to limit the sample to
  a range of wavelengths.