Updated refraction function
After much consideration I finally decided it was time to have a
physically motivated function for the differential refraction of the
atmosphere.

I found a nice paper on the subject by A. Filippenko (1982, PASP 94,
715) which presents a very practical formula as well as a table with
typical values.

Filippenko's equation was implemented as a FittableModel, which is
then used to fit the offsets across the dispersion direction. The only
caveat is that it is necessary to include another parameter, to account
for the field rotation in relation to the parallactic angle.

Also, it is important to remember to convert the offsets from pixels
to arc seconds, or else the parameter limits won't make any sense.