Removed angle from model
The previous version included an additional parameter in the
differential refraction, called cos_angle. The purpose of this
parameter was to allow fitting of the model to individual components
of a rotated reference frame, such as the data cube.

I finally came to the conclusion that this approach is just plain
wrong. Differential refraction should be measured in the data without
reference to the orientation of the reference frame, that is, what
matters is the norm of the offsets.

The new version reflects this reasoning, making no reference to the
image's frame of reference.