Added method to calculate the parallactic angle
The parallactic angle is related to the refraction direction in the
detector, so I thought it would be nice to have the parallactic angle
to check this relation.

Unfortunately the problem is more complex than simply comparing the
two because of the all the other instrumental parameters involved.
I will leave the validation of these angles for a future update.