Skipping already downloaded files
This commit adds a new functionality to gemini_archive.py, which now skips files if it finds that the query list files that are already present in the current directory.

I also took this opportunity to perform a refactoring of the main function, decreasing the number of repeated lines of code, and added some parameters that were missing from the various docstrings.